Refactor node architecture
Only one node is needed.
This node has all the animations in it.
New instances are stored in a dictionary for future access.
When creating a new node, the first parameter is the name of the
node (key in the dictionary) and the second parameter is the animation
name (the sprite-animation stored in AnimatedSprite "Animation").

This way, any node can use any animation.